+++ raytracing/hittable_list.h
@@ -13,7 +13,7 @@ public:
     void add(std::shared_ptr<hittable> object) { objects.push_back(object); }
 
     virtual bool hit(
-        const ray& r, float t_min, float t_max, hit_record& rec) const override;
+        const ray& r, double t_min, double t_max, hit_record& rec) const override;
 
 public:
     std::vector<std::shared_ptr<hittable>> objects;
